Move CriticalInstruction to Reko.Core.Code
