+++ src/Core/Core.csproj
@@ -258,6 +258,7 @@
     <Compile Include="CLanguage\SymbolTable.cs" />
     <Compile Include="CLanguage\TypeSizer.cs" />
     <Compile Include="Code\CodeComment.cs" />
+    <Compile Include="Code\CriticalInstruction.cs" />
     <Compile Include="Code\GotoInstruction.cs" />
     <Compile Include="Code\SwitchInstruction.cs" />
     <Compile Include="Configuration\ArchitectureElement.cs" />
