+++ src/Decompiler/Decompiler.csproj
@@ -117,9 +117,6 @@
       <SubType>Code</SubType>
     </Compile>
     <Compile Include="Analysis\ConstDivisionImplementedByMultiplication.cs" />
-    <Compile Include="Analysis\CriticalInstruction.cs">
-      <SubType>Code</SubType>
-    </Compile>
     <Compile Include="Analysis\DataFlow.cs">
       <SubType>Code</SubType>
     </Compile>
